dkim-verify: Working bh Verification
Finally got the canonicalization for the Body of an email tested and working (seemingly) for hashing. Seems to digest well because it uses an xxd columnation hack to split the hex up accordingly. Then mashes it back together and outputs it in ASCII. I'll be commenting it next before moving on.