Common Functions
Finally got around to adding the common functions library: a place wherein all of the common routines and functions needed by other pieces of the toolkit can source needed generic functions. This helps greatly to trim some of the fat from the rest of the scripts. Moving forward, any function that IS or MIGHT BE shared will end up in the common folder, as a psuedo-library.