DKIM_VERIFY: Complete Overhaul
As promised long ago, I've finally "standardized" the DKIM-verify script to be much more readable. I've added new/clarifying comments, a main method, sections of the main method, modularized a lot of the "main" section's work, fixed some potential problem spots with the && or || booleans, etc etc. I've also finally cleaned up the section for the header signature calculation that I rushed on a while ago and never came back to. Henceforth, I plan on using hard-spaces as indentation, and each nested level of scope will be equivalent to FOUR spaces (instead of the current 2). Still plenty of work to be done here, but I'm sure a nice README update will help to track where we're at on this module.